+++ electron.vite.config.ts
@@ -14,7 +14,8 @@ export default defineConfig({
       rollupOptions: {
         input: {
           index: resolve('src/main/index.ts'),
-          'chart-render-worker': resolve('src/main/modules/shared/chart-render-worker.ts')
+          'chart-render-worker': resolve('src/main/modules/shared/chart-render-worker.ts'),
+          'diagram-render-worker': resolve('src/main/modules/shared/diagram-render-worker.ts')
         }
       }
     }